requirement.txt
certifi==2023.5.7
charset-normalizer==3.1.0
docutils==0.20.1
EasyProcess==1.1
entrypoint2==1.1
futu-api==7.1.3308
holidays==0.27
ibapi==9.81.1.post1
idna==3.4
jeepney==0.8.0
Kivy==2.2.0
Kivy-Garden==0.1.5
KivyCalendar==0.1.3
MouseInfo==0.1.3
mss==9.0.1
numpy==1.25.0
opencv-python==4.8.0.74
pandas==2.0.2
Pillow==9.5.0
protobuf==3.20.3
psutil==5.9.5
PyAutoGUI==0.9.54
pycryptodome==3.18.0
PyGetWindow==0.0.9
Pygments==2.15.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.7
pytz==2023.3
PyYAML==6.0
requests==2.31.0
simplejson==3.19.1
six==1.16.0
tzdata==2023.3
urllib3==2.0.3
